migrate to From<f32>

+++ tessellation/qef.rs
@@ -7,7 +7,7 @@ use std::convert;
 use std::fmt::Debug;
 
 
-pub const EPSILON: f64 = 1e-10;
+pub const EPSILON: f32 = 1e-10;
 
 // Quadratic error function
 
@@ -28,7 +28,7 @@ pub struct Qef<S: 'static + Real + Debug> {
 }
 
 
-impl<S: 'static + Real + Float + Debug + From<f64>> Qef<S> {
+impl<S: 'static + Real + Float + Debug + From<f32>> Qef<S> {
     pub fn new(planes: &[Plane<S>], bbox: BoundingBox<S>) -> Qef<S> {
         let mut qef = Qef {
             solution: na::Vector3::new(S::nan(), S::nan(), S::nan()),
@@ -70,7 +70,7 @@ impl<S: 'static + Real + Float + Debug + From<f64>> Qef<S> {
     pub fn solve(&mut self) {
         let m = &self.ata;
         let ma = na::Matrix3::new(m[0], m[1], m[2], m[1], m[3], m[4], m[2], m[4], m[5]);
-        let sum_as_s: S = convert::From::from(self.num as f64);
+        let sum_as_s: S = convert::From::from(self.num as f32);
         let mean: na::Vector3<S> = self.sum / sum_as_s;
         if let Some(inv) = ma.try_inverse() {
             let b_rel_mean: na::Vector3<S> = self.atb - ma * mean;
@@ -135,7 +135,7 @@ impl<S: 'static + Real + Float + Debug + From<f64>> Qef<S> {
         self.search_solution(accuracy, bbox, ma)
     }
     fn error(&self, point: &na::Vector3<S>, ma: &na::Matrix3<S>) -> S {
-        let _2_as_s: S = convert::From::from(2f64);
+        let _2_as_s: S = convert::From::from(2f32);
         self.btb - _2_as_s * na::dot(point, &self.atb) + na::dot(point, &(*ma * *point))
     }
     pub fn merge(&mut self, other: &Qef<S>) {
